refactor: handleCreateEvent in separate file
