+++ lib/handle-create-event.js
@@ -1,0 +1,48 @@
+module.exports = handleCreateEvent
+
+const readFileSync = require('fs').readFileSync
+const pathResolve = require('path').resolve
+
+const template = readFileSync(pathResolve(__dirname, '..', 'instructions.md'), 'utf8')
+
+function handleCreateEvent (context) {
+  const {ref, refType, repository} = context.payload
+  // run only for newly created branches that start with "first-timers-"
+  if (refType !== 'branch') return
+  if (!/^first-timers-/.test(ref)) return
+
+  const api = context.github
+  const owner = repository.owner.login
+  const repo = repository.name
+
+  api.repos.getBranch({
+    owner: owner,
+    repo: repo,
+    branch: ref
+  })
+
+  .then(function (result) {
+    const sha = result.data.commit.sha
+    return api.repos.getCommit({
+      owner: owner,
+      repo: repo,
+      sha: sha
+    })
+  })
+
+  .then(function (result) {
+    const commitMessage = result.data.commit.message
+    const {filename, patch, blob_url} = result.data.files[0]
+    const content = template.replace(/\$DIFF/, patch).replace(/\$FILENAME/, filename).replace(/\$BLOB_URL/, blob_url).replace(/\$REPO/, repo)
+
+    return api.issues.create({
+      owner: owner,
+      repo: repo,
+      title: commitMessage,
+      body: content,
+      labels: ['test-newbie']
+    })
+  })
+
+  .catch(console.log.bind(console))
+}
